SaveProgram class, update issues and search id
I’ve added a class for saving the current state of the program to and
object file on the system. Added another function with buttons which
updates an issue depending on the user input.
And I’ve also added a simple search function to find specific issues
using ID as a parameter.
Added a Menu which is not currently working. The save methods are also
a bit buggy. Hopefully they will be happy with this result anyway.